Move Vis class to pithy.web.asgi module
